# HabitFlow.Api/openapi.json
{
...
  "openapi": "3.0.0",
  "info": {
    "title": "HabitFlow API",
    "description": "REST API for HabitFlow habit tracking application",
    "version": "v1"
  },
  "paths": {
...
    "/api/v1/habits": {
      "get": {
        "tags": [
          "Habits"
        ],
        "operationId": "GetHabits",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "nullable": true
            },
            "x-position": 1
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "nullable": true
            },
            "x-position": 2
          },
          {
            "name": "type",
            "in": "query",
            "schema": {
              "nullable": true,
              "oneOf": [
                {
                  "$ref": "#/components/schemas/HabitType"
                }
              ]
            },
            "x-position": 3
          },
          {
            "name": "completionMode",
            "in": "query",
            "schema": {
              "nullable": true,
              "oneOf": [
                {
                  "$ref": "#/components/schemas/CompletionMode"
                }
              ]
            },
            "x-position": 4
          },
          {
            "name": "active",
            "in": "query",
            "schema": {
              "type": "boolean",
              "nullable": true
            },
            "x-position": 5
          },
          {
            "name": "search",
            "in": "query",
            "schema": {
              "type": "string",
              "nullable": true
            },
            "x-position": 6
          },
          {
            "name": "sortField",
            "in": "query",
            "schema": {
              "nullable": true,
              "oneOf": [
                {
                  "$ref": "#/components/schemas/HabitSortField"
                }
              ]
            },
            "x-position": 7
          },
          {
            "name": "sortDirection",
            "in": "query",
            "schema": {
              "nullable": true,
              "oneOf": [
                {
                  "$ref": "#/components/schemas/SortDirection"
                }
              ]
            },
            "x-position": 8
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedResponseOfHabitResponse"
                }
              }
            }
          },
          "401": {
            "description": ""
          }
        }
      },
      "post": {
        "tags": [
          "Habits"
        ],
        "operationId": "CreateHabit",
        "requestBody": {
          "x-name": "request",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateHabitRequest"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "400": {
            "description": ""
          },
          "401": {
            "description": ""
          },
          "409": {
            "description": ""
          }
        }
      }
    },
    "/api/v1/habits/{id}": {
      "get": {
        "tags": [
          "Habits"
        ],
        "operationId": "GetHabit",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HabitResponse"
                }
              }
            }
          },
          "401": {
            "description": ""
          },
          "404": {
            "description": ""
          }
        }
      },
      "patch": {
        "tags": [
          "Habits"
        ],
        "operationId": "UpdateHabit",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 1
          }
        ],
        "requestBody": {
          "x-name": "request",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateHabitRequest"
              }
            }
          },
          "required": true,
          "x-position": 2
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "400": {
            "description": ""
          },
          "401": {
            "description": ""
          },
          "404": {
            "description": ""
          }
        }
      },
      "delete": {
        "tags": [
          "Habits"
        ],
        "operationId": "DeleteHabit",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "204": {
            "description": ""
          },
          "401": {
            "description": ""
          },
          "404": {
            "description": ""
          }
        }
      }
    },
    "/api/v1/habits/{id}/calendar": {
      "get": {
        "tags": [
          "Habits"
        ],
        "operationId": "GetHabitCalendar",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 1
          },
          {
            "name": "from",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            },
            "x-position": 2
          },
          {
            "name": "to",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            },
            "x-position": 3
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HabitCalendarResponse"
                }
              }
            }
          },
          "400": {
            "description": ""
          },
          "401": {
            "description": ""
          },
          "404": {
            "description": ""
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "PagedResponseOfHabitResponse": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "totalCount": {
            "type": "integer",
            "format": "int32"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/HabitResponse"
            }
          }
        }
      },
      "HabitResponse": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "type": {
            "$ref": "#/components/schemas/HabitType"
          },
          "completionMode": {
            "$ref": "#/components/schemas/CompletionMode"
          },
          "daysOfWeekMask": {
            "type": "integer",
            "format": "int32"
          },
          "targetValue": {
            "type": "integer",
            "format": "int32"
          },
          "targetUnit": {
            "type": "string",
            "nullable": true
          },
          "deadlineDate": {
            "type": "string",
            "format": "date",
            "nullable": true
          },
          "createdAtUtc": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "HabitType": {
        "type": "string",
        "description": "",
        "x-enumNames": [
          "Start",
          "Stop"
        ],
        "enum": [
          "Start",
          "Stop"
        ]
      },
      "CompletionMode": {
        "type": "string",
        "description": "",
        "x-enumNames": [
          "Binary",
          "Quantitative"
        ],
        "enum": [
          "Binary",
          "Quantitative"
        ]
      },
      "HabitSortField": {
        "type": "string",
        "description": "",
        "x-enumNames": [
          "CreatedAtUtc",
          "Title",
          "DeadlineDate"
        ],
        "enum": [
          "CreatedAtUtc",
          "Title",
          "DeadlineDate"
        ]
      },
      "SortDirection": {
        "type": "string",
        "description": "",
        "x-enumNames": [
          "Asc",
          "Desc"
        ],
        "enum": [
          "Asc",
          "Desc"
        ]
      },
      "CreateHabitRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "type": {
            "$ref": "#/components/schemas/HabitType"
          },
          "completionMode": {
            "$ref": "#/components/schemas/CompletionMode"
          },
          "daysOfWeekMask": {
            "type": "integer",
            "format": "byte"
          },
          "targetValue": {
            "type": "integer"
          },
          "targetUnit": {
            "type": "string",
            "nullable": true
          },
          "deadlineDate": {
            "type": "string",
            "format": "date",
            "nullable": true
          }
        }
      },
      "UpdateHabitRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "title": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "type": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/HabitType"
              }
            ]
          },
          "completionMode": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/CompletionMode"
              }
            ]
          },
          "daysOfWeekMask": {
            "type": "integer",
            "format": "byte",
            "nullable": true
          },
          "targetValue": {
            "type": "integer",
            "nullable": true
          },
          "targetUnit": {
            "type": "string",
            "nullable": true
          },
          "deadlineDate": {
            "type": "string",
            "format": "date",
            "nullable": true
          },
          "clearDeadline": {
            "type": "boolean",
            "nullable": true
          }
        }
      },
      "HabitCalendarResponse": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "habitId": {
            "type": "integer",
            "format": "int32"
          },
          "from": {
            "type": "string",
            "format": "date"
          },
          "to": {
            "type": "string",
            "format": "date"
          },
          "days": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/HabitCalendarDay"
            }
          }
        }
      },
      "HabitCalendarDay": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "date": {
            "type": "string",
            "format": "date"
          },
          "isPlanned": {
            "type": "boolean"
          },
          "actualValue": {
            "type": "integer",
            "format": "int32"
          },
          "targetValueSnapshot": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "completionModeSnapshot": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/CompletionMode"
              }
            ]
          },
          "habitTypeSnapshot": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/HabitType"
              }
            ]
          },
          "dailyScore": {
            "type": "number",
            "format": "double"
          }
        }
      },
...
    }
  }
}